+++ src/modals/FrameViewerModal/index.tsx
@@ -3,22 +3,25 @@ import { forwardRef, useState } from 'react';
 import { Modal } from '@/components/Modal.js';
 import { useSingletonModal } from '@/hooks/useSingletonModal.js';
 import type { SingletonModalRefCreator } from '@/libs/SingletonModal.js';
-import { FrameViewer } from '@/modals/FrameV2Modal/FrameViewer.js';
+import { FrameViewer } from '@/modals/FrameViewerModal/FrameViewer.js';
 import type { FrameV2 } from '@/types/frame.js';
 
-export type FrameV2ModalProps = {
+export type FrameViewerModalOpenProps = {
     frame: FrameV2;
 };
-export type FrameV2ModalCloseProps = void;
+export type FrameViewerModalCloseProps = void;
 
-export const FrameV2Modal = forwardRef<SingletonModalRefCreator<FrameV2ModalProps>>(
-    function FrameV2Modal(_, ref) {
-        const [props, setProps] = useState<FrameV2ModalProps>();
+export const FrameViewerModal = forwardRef<SingletonModalRefCreator<FrameViewerModalOpenProps, FrameViewerModalCloseProps>>(
+    function FrameViewerModal(_, ref) {
+        const [props, setProps] = useState<FrameViewerModalOpenProps>();
 
         const [open, dispatch] = useSingletonModal(ref, {
             onOpen(p) {
                 setProps(p);
             },
+            onClose() {
+                setProps(undefined);
+            }
         });
 
         const onClose = () => dispatch?.close();
